refactor(calculator): move logic related to tokenize to StringExpressionTokenizer
